tests: Fix unordered comparison of lists.
As the order of items from FilterSet.filter_set is nondeterministic,
it is possible that the sequence for string placeholders for the query
string generated by the filter helper might be different, leading to test
failures, while the order within the set doesn't actually matter. This
commit fixes this problem by using assertCountEqual instead for list
comparison.